Extract reusable `Timeable` concern and apply it to factories
