+++ database/factories/IncomeFactory.php
@@ -7,11 +7,14 @@ use App\Models\Category;
 use App\Models\Income;
 use App\Models\Person;
 use App\Models\User;
+use Database\Factories\Concerns\Timeable;
 use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Carbon;
 
 class IncomeFactory extends Factory
 {
+    use Timeable;
+
     protected $model = Income::class;
 
     public function definition(): array
@@ -28,26 +31,5 @@ class IncomeFactory extends Factory
             'account_id' => Account::factory(),
             'category_id' => Category::factory(),
         ];
-    }
-
-    public function today(): self
-    {
-        return $this->state(fn (array $attributes) => [
-            'transacted_at' => Carbon::now(),
-        ]);
-    }
-
-    public function tomorrow(): self
-    {
-        return $this->state(fn (array $attributes) => [
-            'transacted_at' => Carbon::now()->addDay(),
-        ]);
-    }
-
-    public function yesterday(): self
-    {
-        return $this->state(fn (array $attributes) => [
-            'transacted_at' => Carbon::now()->subDay(),
-        ]);
     }
 }
